fix: resolved MIME type issue for CSS file
Fixed an issue where the browser refused to apply the stylesheet due to an incorrect MIME type ('text/html').

Error message encountered:
"Refused to apply style from 'http://127.0.0.1:5500/css/styles.css' because its MIME type ('text/html') is not a supported stylesheet MIME type, and strict MIME checking is enabled."

The problem was due to a missing the first backslash  and i resolved it by adding it in the href  /css/style.css.